refactor: modularize 4D projection
